+++ Changelog.md
@@ -30,6 +30,7 @@ Rust:
 - add plugin types
 - update validity checks on memory management
 - intermingle gates and function declarations into a list of directives
+- rename Public/PrivateInput gates by Public/Private
 
 # SIEVE IR v1.0.1, 2022-08